requirements.txt
#required python packages for ob1 arm repo
absl-py==0.11.0
actionlib==1.13.2
aiohttp==3.8.3
aiosignal==1.3.1
angles==1.9.13
async-timeout==4.0.2
attrs==22.1.0
bondpy==1.8.6
camera-calibration==1.16.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==0.5.2
charset-normalizer==2.1.1
contourpy==1.0.6
controller-manager==0.19.5
controller-manager-msgs==0.19.5
cv-bridge==1.16.2
cycler==0.11.0
Cython==0.29.32
defusedxml==0.7.1
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.8.0
docutils==0.19
dynamic-reconfigure==1.7.3
empy==3.3.4
exceptiongroup==1.0.4
fcl==0.0.12
fonttools==4.38.0
frozenlist==1.3.3
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
h5py==3.7.0
idna==3.4
image-geometry==1.16.2
importlib-metadata==5.0.0
iniconfig==1.1.1
interactive-markers==1.12.0
joblib==1.2.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
kdl-parser-py==1.14.2
keras==2.11.0
kinpy==0.2.0
kiwisolver==1.4.4
laser_geometry==1.6.7
llvmlite==0.39.1
lxml==4.9.1
matplotlib==3.6.2
message-filters==1.15.14
moveit-commander==1.1.10
moveit-core==1.1.10
moveit-ros-planning-interface==1.1.10
moveit-ros-visualization==1.1.10
msgpack==1.0.4
multidict==6.0.2
multitables==2.0.1
netifaces==0.11.0
numba==0.56.4
numexpr==2.8.4
numpy==1.23.4
orjson==3.8.1
packaging==21.3
Pillow==9.3.0
pluggy==1.0.0
ply==3.11
pyaml==21.10.1
pyassimp==4.1.4
pyparsing==3.0.9
pyquaternion==0.9.9
pytest==7.2.0
pytest-repeat==0.9.1
python-dateutil==2.8.2
python-fcl==0.7.0.1
python-qt-binding==0.4.4
PyYAML==5.4.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
resource_retriever==1.12.7
rosbag==1.15.14
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.15.14
roslaunch==1.15.14
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.14
rosmake==1.15.8
rosmaster==1.15.14
rosmsg==1.15.14
rosnode==1.15.14
rosparam==1.15.14
rospkg==1.4.0
rospy==1.15.14
rosservice==1.15.14
rostest==1.15.14
rostopic==1.15.14
rosunit==1.15.8
roswtf==1.15.14
rqt-moveit==0.5.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.14
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.11
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_image_view==0.4.16
rqt_joint_trajectory_controller==0.20.0
rqt_launch==0.4.9
rqt_logger_level==0.4.11
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_runtime_monitor==0.5.9
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_tf_tree==0.6.3
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rviz==1.14.19
scikit-learn==1.1.3
scipy==1.9.3
sensor-msgs==1.13.1
sip==6.7.4
six==1.16.0
sklearn==0.0.post1
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
srdfdom==0.6.3
tables==3.7.0
tdqm==0.0.1
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
topic-tools==1.15.14
tqdm==4.64.1
transformations==2020.1.1
urdfdom-py==0.4.6
vtk==9.2.2
wrapt==1.14.1
wslink==1.9.1
xacro==1.14.13
yarl==1.8.1
zipp==3.10.0
